+++ src/main/java/ai/starlake/transpiler/schema/treebuilder/TreeBuilder.java
@@ -1,5 +1,6 @@
 package ai.starlake.transpiler.schema.treebuilder;
 
+import ai.starlake.transpiler.JSQLColumResolver;
 import ai.starlake.transpiler.schema.JdbcResultSetMetaData;
 
 import java.sql.SQLException;
@@ -11,5 +12,5 @@ public abstract class TreeBuilder<T> {
     this.resultSetMetaData = resultSetMetaData;
   }
 
-  public abstract T getConvertedTree() throws SQLException;
+  public abstract T getConvertedTree(JSQLColumResolver resolver) throws SQLException;
 }